Created jajanken.js
Version 2: I externally linked my JS code and got the logic of the program working correctly as well as found a way to keep count of the score by changing the return values of the "winner" function. In the future, I will revisit this project to add a Hunter x Hunter themed GUI so it can truly be a game of Jajanken Rock-Paper-Scissors!