Aesthetics
I worked immensely on making the game look pretty and added images which represent the choice made each round by the player and computer. I need to continue fixing the removeEventListener method on the buttons so the game properly ends when a score of 5 is reached. I also need to continue working more on the aesthetics